move language utils to display.utils folder
